% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_regression.R
\name{lasso_regression}
\alias{lasso_regression}
\title{Lasso Regression}
\usage{
lasso_regression(
  x,
  y,
  alpha = 1,
  lambda = NULL,
  importance = FALSE,
  type = "default",
  nfolds = 10,
  ignoreWarnings = T,
  k = 6
)
}
\arguments{
\item{x}{A data frame or matrix containing the predictor variables.}

\item{y}{A numeric vector or factor for regression and classification tasks respectively.}

\item{alpha}{A numeric value between 0 and 1. It controls the elastic net mixing parameter. If alpha = 1, the function performs Lasso regression.}

\item{lambda}{A numeric value for the regularization parameter. If not provided, it defaults to 0.01.}

\item{importance}{A logical value. If TRUE, the function returns the importance of the predictors.}

\item{type}{A character string. It can be either 'default' for regression tasks or 'class' for classification tasks.}

\item{nfolds}{An integer value. It is used for cross-validation in the case of importance = TRUE and k is not provided.}

\item{ignoreWarnings}{A logical value. If TRUE, warnings are ignored.}

\item{k}{An integer value. It is used to select the top k most important predictors.}
}
\value{
A list containing the glmnet fit, coefficients, and selected features.
}
\description{
This function performs Lasso Regression using the glmnet package. It allows for both regression and classification tasks.
}
\examples{
# Load the required library
library(glmnet)
# Generate sample data
set.seed(123)
x <- matrix(rnorm(100), ncol = 5)
y <- rnorm(20)
# Fit a Lasso regression model
lasso_model <- lasso_regression(x = x, y = y, alpha = 1, lambda = NULL, importance = FALSE, type = "default", nfolds = 10, ignoreWarnings = TRUE, k = 3)
# Print the selected features
print(lasso_model$selectedFeatures)
# Print the coefficients
print(lasso_model$coef)
# Print the fitted model
print(lasso_model$fit)
# Make predictions if needed
#  predictions <- predict(lasso_model$fit, newx = x_test)
}
