% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear_regression}
\alias{linear_regression}
\title{Linear Regression}
\usage{
linear_regression(
  x,
  y,
  alpha = 1,
  lambda = NULL,
  importance = FALSE,
  type = "default",
  nfolds = 10,
  ignoreWarnings = T,
  k = 6
)
}
\arguments{
\item{x}{A matrix or data frame of predictor variables.}

\item{y}{A numeric vector of response variable.}

\item{alpha}{The elastic-net mixing parameter. Default is 1 (ridge regression).}

\item{lambda}{The value of the penalty parameter lambda. If NULL, a default value of 0.01 is used.}

\item{importance}{Logical. If TRUE, the function will calculate feature importance using the k important predictors.}

\item{type}{Character string indicating the type of model. Currently, only "default" is supported.}

\item{nfolds}{Number of folds for cross-validation. Default is 10.}

\item{ignoreWarnings}{Logical. If TRUE, warnings will be suppressed. Default is TRUE.}

\item{k}{The number of important predictors to select if importance is TRUE. Default is 6.}
}
\value{
A list containing the fitted model, coefficients, and selected features.
}
\description{
This function performs linear regression using the glmnet package.
}
\examples{
x <- matrix(rnorm(100*20), 100, 20)
y <- rnorm(100)
fit <- linear_regression(x, y)
summary(fit$fit)
coef(fit$fit)
fit$selectedFeatures
}
