% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_regression}
\alias{predict_regression}
\title{Predict Regression}
\usage{
predict_regression(coefficients, newdata, fit = NULL)
}
\arguments{
\item{coefficients}{A matrix of regression coefficients.}

\item{newdata}{The new data for prediction.}

\item{fit}{The fitted model object (optional).}
}
\value{
A vector of predicted values.
}
\description{
Predicts the response variable using the coefficients and new data.
}
